Extend targets: Android, JVM for most part, and nominal iOS bits.
This comes with trickly KMM build structure that hacks around plugin
resolution (see `setttings.gradle`).

Most of the existing tests are moved back to `commonMain` so that they can be
tested on JS, native and Android as well.
This still leaves `MidiPlayerTest` out of the loop, as it depends on
`runBlocking{...}` which is JVM-only.

`AlsaMidiAccess` is still part of JVM as it depends on alsakt which does not
build on non-Linux platform and we need some working package provider that
hosts alsakt.