Rename artifactId: "ktmidi-kotlinMultiplatform" (default) to "ktmidi"
Why? Because otherwise it breaks GitHub Actions packaging!

There are some predecessors:
https://github.community/t/gradle-maven-deploy-failing-with-422-unprocessable-entity/137299

> artifacId should be lowercase to be able to push to github packages

It was surprising to me, because, "{name}-kotlinMultiplatform" is the *default*
artifactId that Kotlin MPP itself gives us. Is that even a Maven restriction,
or GitHub Actions restriction? Let's see...

https://maven.apache.org/guides/mini/guide-naming-conventions.html

> **artifactId** is the name of the jar without version. If you created it, then you can choose whatever name you want with lowercase letters and no strange symbols.

So, GitHub Actions is not faulty. It is a fault at Kotlin MPP itself.

¯\_(ツ)_/¯

In any case, no we have to rename everything that is based on this naming
convention...