reorganize the project into two projects.
As Kotlin 1.6 really breaks "MPP + some module" kind of projects (as it
happened to augene-ng and mugene-ng too) ktmidi-jvm-desktop has to be
kicked out of ktmidi project. This MPP limitation is super annoying and
super error prone, easy to bring in bugs due to version mismatches, but
we have no remedy.

As a result, there are now ktmidi-project and extension-project. Any new
future packages would go into extension-project.