added translateUmpToMidi1Bytes() and renamed UmpTranslatorContext.
`UmpTranslatorContext` is now confusing as it could sound applicable
to both UMP to MIDI1 and MIDI1 to UMP. The existing one only applies
to the latter, so renamed it to `Midi1ToUmpTranslatorContext`.

The new `translateUmpToMidi1Bytes()` tries to convert Midi2Music
`Ump` stream to SMF bytes. Note that this mentions Midi2Music and SMF,
not UMP and MIDI1 bytestream, because they involve delta times.
For more details, see the code comment on the new function.