unify ktmidi-project and extension-project again.
It was split because gradle could not handle multiple multiplatform prorjects.
I guess it is still the case, but now that we were rather suffered from the
version maintainability glitch, it would make better sense to keep them in
the same project as long as possible. extension-project is not multiplatform,
so it is okay to have it unified. When we want to create another MPP module,
we would split things again.