go back to "resolve `implementation` via maven" instead of `project()`.
JavaCPP brings in another complication that `XYZ-platform` does not exist as a
module but created as part of `XYZ` module. There is no way to resolve the
`-platform` jar.

We may have to fight the Gradle package resolution failure issue if it still
happens, but now we removed all those platforms jars from `rtmidi-javacpp`
things may work now...