[ci,citool] start implementing MIDI-CI client (Initiator) part.
Mixing Initiator and Responder via the same virtual port is not clean, but
it seems we cannot simply create multiple virtual ports on the current
rtmidi backend and/or wrapper, so we switch the uses of them by the main UI tab.

It also involves client-side catch-ups; CIFactory started returning the CI
buffer slices, and event handler model is taking place (same as responder).