[ci,citool] Pair(MidiCIProfileId,Boolean) is now MidiCIPrpfile, with "address".
Addressing is also an important information on a Profile state, so we keep
those 3 items in a new MidiProfile data class.

We could name it as MidiCIProfileState, but that would bring confusion in
Compose MPP app API surface (it is not a State<T>).
MidiCIProfile is similarly confusing (it should indicate that it is a dynamic
state, not anything static) so we might rename it later...