[citool] un-statify MIDI device lists.
They are queried every time device list dropdown shows up, and that's already
what we want. No need to make them state list and update every time.

They rather seem to have caused "Reading a state that was created after the
snapshot was taken or in a snapshot that has not yet been applied" crashers.