[ci,citool] change mime type default policy. Use it for better property editing
It is Common Rules' responsibility to fill non-null mime type now.

It was kind of needed due to weird null value handling in Jetpack Compose
that fails to provide null value state for the mime type.
The way how we show the editor/textview for property value is more obvious
on why they can be shown and editable.