[ci,citool] More MidiCIResponder profile configuration implementation.
There was a notable bug in JUCE that incorrectly expects "Number Channels
Requested" field in Set Profile On message as always non-zero, whereas the
specification explicitly states it is 0 when the address is 0x7E or 0x7F.

Since it is annoying to not have a workaround, now we have a new settings
screen (that was left as TODO) and has an option to always send 1 even for
those addresses. With that option, the InitiatorModel behaves as such.