[ci,citool] bring back tests to work and fixed Mcoded7 support.
The test sources were placed in the wrong directory and thus they had been
actually broken and not run(!) until now.

Mcoded7 implementation is fixed to interoperate enough with juce_midi_ci
(it was implemented as if it operated over bytes not bits)

The UI parts for mutualEncoding selector has been adjusted to make sense.