MIDI-CI: reduced various Common Rules for PE specific strongly-typed APIs.
MidiCIInitiator and MidiCIResponder are supposed to not be too tied to
Common Rules for PE, but in reality there has been a lot of strongly-typed
functions and classes that cannot live without that specification.
That needs to be fixed, if not all. This removes a handful of such API parts.