a bunch of MIDI Message Report related fixes to make it functional.
Now that we have a working initiator and a responder (both ktmidi-ci-tool),
we can test MIDI Message Report functionality. And it was not working due to
various issues.
Now they are fixed, and MIDI messages are logged in the manner that makes sense.

Formerly it was logging every single MIDI input to the logger which resulted
in hundreds of lines (and if it were MIDI2 things could go even worse).
Now those MIDI report bytes are logged in per-channel chunk, so it is 16
for MIDI1 device. (For MIDI2 it could be 256 in the worst case; by the time
we support MIDI2 logging we would add some filtering options...)