[ci,citool] restructured how property values are stored. Implemented pagination.
Pagination is specified at Common Rules for PE section 6.6.2.

Local property values are stored at both MidiCIResponder and
CommonRulesPropertyService, and that was causing inconsistency issue. Also,
the latter saved property values as in JSON, which drops support for non-JSON
property values. They are now covered (instead we parse JSON every time, but
it would not matter... JSON is not RT safe at least).