MIDI-CI: implemented `serialize()` methods in most of the Message classes.
It is so far not an abstract method that every class needs to implement.
There are classes that do not implement it. And Some property exchange
messages return `Sequence<List<Byte>`` instead of `List<Byte>` as they can
span more than one message chunks. We don't know how we would deal with the
difference.