API-BREAKING: start v0.8.0 development with MIDI 2.0 device access. [WIP]
It is in reflection to Android 15 virtual MIDI device support.

This involves API breakage in MidiAccess API: MidiPort does not have
midiProtocol anymore (it was not designed to work with June 2023 updates).
We instead have midiTransportProtocol in MidiPortDetails.

All those references to MidiCIProtocolType constants are removed.

I still do not see the MIDI 2.0 virtual devices showing up yet, so it is
likely a WIP. ktmidi-ci-tool is implemented to show MIDI 2.0 transport devices.