ci,citool: more MidiCIDevice refactoring. We have PC/PE facades now.
Those new classes provides end-user developer Profile Configuration and
Property Exchange features (PE facade is still messed but we will sort out).

PropertyExchangeResponder is gone as part of the refactoring. Some of the
moved members to the facade may go back into Messenger. But we'd rather
have to figure out which features should be moved out from
CommonRulesPropertyService to the new facade first.